+++ CHANGELOG.md
@@ -36,7 +36,7 @@ Before releasing:
 
 - Implemented a conversion between `bool` and `LogicLevel`. (#428)
 - Added support for capturing backtraces without heap allocations via the `BacktraceIter` struct. (#456)
-- Added `DistanceSensor::ready` which returns a future resolving when the sensor is ready to be used. (#462)
+- Added `DistanceSensor::wait_ready` which returns a future resolving when the sensor is ready to be used. (#462)
 
 ### Fixed